add unit test + start changing the metadata methods to streaming versions
